Created a fragment and activity corresponding to the login layout.
Created a login container.

loginActivity is now the main and launcher activity.

Updated some comments and strings which still referred to band.
            Added some strings for the new layouts to use.

Created a prototype validateLogin method to be called by the submit button on the login fragment. This is and related method(s) are commented for now due to an error.

Created a register layout and connected it to a fragment that appears when clicking the register button on the login screen.

Register fragment properly adds the user's username and password into the database and returns the user to the login screen.